+++ kwcoco/cli/__main__.py
@@ -12,6 +12,7 @@ def main(cmdline=True, **kw):
     from kwcoco.cli import coco_split
     from kwcoco.cli import coco_show
     from kwcoco.cli import coco_toydata
+    # from kwcoco.cli import coco_rebase
 
     # Create a list of all submodules with CLI interfaces
     cli_modules = [
@@ -19,6 +20,7 @@ def main(cmdline=True, **kw):
         coco_union,
         coco_split,
         coco_show,
+        # coco_rebase,
         coco_toydata,
     ]
 
@@ -26,7 +28,8 @@ def main(cmdline=True, **kw):
     # the previous CLI interfaces.
     parser = argparse.ArgumentParser(
         description='The Kitware COCO CLI',
-        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+        # formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+        formatter_class=argparse.RawDescriptionHelpFormatter,
     )
     subparsers = parser.add_subparsers(help='specify a command to run')
 
@@ -34,8 +37,17 @@ def main(cmdline=True, **kw):
         cli_cls = cli_module._CLI
         subconfig = cli_cls.CLIConfig()
 
-        subparser = subparsers.add_parser(
-                cli_cls.name, help=subconfig.__class__.__doc__)
+        # TODO: make subparser.add_parser args consistent with what
+        # scriptconfig generates when parser=None
+        if hasattr(subconfig, '_parserkw'):
+            parserkw = subconfig._parserkw()
+        else:
+            # for older versions of scriptconfig
+            parserkw = dict(
+                description=subconfig.__class__.__doc__
+            )
+        parserkw['help'] = parserkw['description']
+        subparser = subparsers.add_parser(cli_cls.name, **parserkw)
         subparser = subconfig.argparse(subparser)
         subparser.set_defaults(main=cli_cls.main)
 
@@ -53,6 +65,7 @@ def main(cmdline=True, **kw):
         ret = main(cmdline=False, **kw)
     except Exception as ex:
         print('ERROR ex = {!r}'.format(ex))
+        raise
         return 1
     else:
         if ret is None: