+++ kwcoco/util/__init__.py
@@ -1,0 +1,12 @@
+"""
+mkinit ~/code/kwcoco/kwcoco/util/__init__.py -w
+"""
+
+from kwcoco.util import util_futures
+from kwcoco.util import util_sklearn
+
+from kwcoco.util.util_futures import (Executor, SerialExecutor,)
+from kwcoco.util.util_sklearn import (StratifiedGroupKFold,)
+
+__all__ = ['Executor', 'SerialExecutor', 'StratifiedGroupKFold',
+           'util_futures', 'util_sklearn']